requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
redis<4.0.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# Web search and scraping
duckduckgo-search>=3.9.0
wikipedia>=1.4.0
ollama>=0.1.0

# LLM and AI
transformers>=4.36.0
torch>=2.0.0
sentencepiece>=0.1.99
protobuf>=4.25.1
sentence-transformers>=2.2.2

# Data processing and analysis
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Dataset APIs
kaggle>=1.6.6
huggingface-hub>=0.20.3
PyGithub>=2.1.1

# Caching and performance
redis-py-cluster>=2.1.3
aioredis>=2.0.1

# Logging and monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Testing and validation
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Development tools
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0

# Web framework (for API)
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.4.0

# UI framework
streamlit>=1.24.0

# Notifications
slack-sdk>=3.21.0

# Added from the code block
fake-useragent>=1.1.0
